requirements.txt
openai
cn2an
qq-botpy
requests
PyYAML
pydantic
pilk